Always increasing counter approach.
Since we check for existing directories prior to increasing the counter,
we could get a directory which was previously used(deleted meanwhile)
which could cause unexpected behaviors.

We now store the paths we're trying to create in a dictionary and that's
what will be used as the source for the counter.